Deleted the hardcoded 'merida' search test from the build command.
Added a load() method to the InvertedIndex class. It should load the index and docmap from disk using the pickle module's load function

Updated the search command to load the index from disk. If it doesn't exist, just print an error message and exit.

Removed the logic for iterating over queries and movies.
Instead, iterate over each token in the query, and use the inverted index to get any matching documents for each token