adding the term frequency to the project
Add a new term_frequencies attribute to your InvertedIndex class. It's a
dictionary of document IDs to Counter objects. A Counter is a dictionary
optimized for counting. We'll be keeping track of how many times each
term appears in each document.
Update your load and save methods to also load and save this new
attribute into a new cache/term_frequencies.pkl file.
Update your __add_document method to update the term frequencies for
each token in the document. For each token, increment its count in the
Counter for that document ID.
Add a new get_tf(self, doc_id, term) method. It should return the times
the token appears in the document with the given ID. If the term doesn't
exist in that document, return 0. Be sure to tokenize the term, but
assume that there is only one token. If there's more than one, raise an
exception.
Add a new "tf" command to the keyword_search_cli.py entrypoint script.

    takes a document ID and a term as arguments.
    prints the term frequency for that term in the document with the
given ID.